# tomford/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "User API",
        "description": "APIs for user management",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/product/create": {
            "post": {
                "tags": ["product-api-controller"],
                "operationId": "createProduct",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProductCommand"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "integer",
                                    "format": "int32"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/product-option/create": {
            "post": {
                "tags": ["product-option-api-controller"],
                "operationId": "createProduct_1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProductOptionCommand"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/product-image/create": {
            "post": {
                "tags": ["product-image-api-controller"],
                "operationId": "createProduct_2",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProductImageCommand"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/order/create": {
            "post": {
                "tags": ["order-api-controller"],
                "operationId": "createOrder",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrderCommand"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/product/{productId}": {
            "get": {
                "tags": ["product-api-controller"],
                "operationId": "getProductById",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductDetailDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateProductCommand": {
                "required": ["categoryId", "description", "displayImage", "name", "price"],
                "type": "object",
                "properties": {
                    "name": {
                        "maxLength": 100,
                        "minLength": 5,
                        "type": "string"
                    },
                    "forGender": {
                        "type": "string",
                        "enum": ["FOR_MALE", "FOR_FEMALE", "FOR_BOTH"]
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "minimum": 1000,
                        "type": "number",
                        "format": "double"
                    },
                    "discount": {
                        "minimum": 0,
                        "type": "number",
                        "format": "double"
                    },
                    "displayImage": {
                        "type": "string"
                    },
                    "brandId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "categoryId": {
                        "minimum": 1,
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "CreateProductOptionCommand": {
                "required": ["colorId", "productId", "size"],
                "type": "object",
                "properties": {
                    "colorId": {
                        "minimum": 1,
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "string"
                    },
                    "stock": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "productId": {
                        "minimum": 1,
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "CreateProductImageCommand": {
                "type": "object",
                "properties": {
                    "productId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "colorId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "url": {
                        "type": "string"
                    }
                }
            },
...
            "CreateOrderCommand": {
                "required": ["address", "customerName", "note", "paymentMethod", "phoneNumber", "productOptionIds"],
                "type": "object",
                "properties": {
                    "productOptionIds": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    "customerName": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    },
                    "note": {
                        "type": "string"
                    },
                    "promotionCode": {
                        "type": "string"
                    },
                    "paymentMethod": {
                        "type": "string",
                        "enum": ["COD", "MOMO_QR", "MOMO_ATM"]
                    }
                }
            },
...
            "CategoryBriefDto": {
                "type": "object",
                "properties": {
                    "createdBy": {
                        "type": "string"
                    },
                    "createdDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "lastModifiedBy": {
                        "type": "string"
                    },
                    "lastModifiedDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "categoryId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "parent": {
                        "$ref": "#/components/schemas/CategoryBriefDto"
                    }
                }
            },
            "ColorDto": {
                "type": "object",
                "properties": {
                    "colorId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "name": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    }
                }
            },
            "ProductDetailDto": {
                "type": "object",
                "properties": {
                    "createdBy": {
                        "type": "string"
                    },
                    "createdDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "lastModifiedBy": {
                        "type": "string"
                    },
                    "lastModifiedDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "productId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "name": {
                        "type": "string"
                    },
                    "forGender": {
                        "type": "string",
                        "enum": ["FOR_MALE", "FOR_FEMALE", "FOR_BOTH"]
                    },
                    "slug": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number",
                        "format": "double"
                    },
                    "discount": {
                        "type": "number",
                        "format": "double"
                    },
                    "displayImage": {
                        "type": "string"
                    },
                    "category": {
                        "$ref": "#/components/schemas/CategoryBriefDto"
                    },
                    "productOptions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductOptionDto"
                        }
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductImageDto"
                        }
                    },
                    "description": {
                        "type": "string"
                    }
                }
            },
            "ProductImageDto": {
                "type": "object",
                "properties": {
                    "url": {
                        "type": "string"
                    },
                    "forColor": {
                        "$ref": "#/components/schemas/ColorDto"
                    }
                }
            },
            "ProductOptionDto": {
                "type": "object",
                "properties": {
                    "createdBy": {
                        "type": "string"
                    },
                    "createdDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "lastModifiedBy": {
                        "type": "string"
                    },
                    "lastModifiedDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "productOptionId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "string"
                    },
                    "stock": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "deletedDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "color": {
                        "$ref": "#/components/schemas/ColorDto"
                    }
                }
            },
...
        }
    }
}